backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
gunicorn==21.2.0
python-dotenv==1.0.0
aiohttp==3.8.5
motor==3.3.1
redis==4.4.0
aioredis==2.0.1
tenacity==8.2.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
uvloop==0.17.0
httptools==0.6.0
prometheus-fastapi-instrumentator==6.1.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-jaeger==1.22.0
python-json-logger==2.0.7
pytest==7.4.3
pytest-cov==4.1.0
mypy==1.7.1
